requirements.txt
toml==0.9.2
discord.py==0.16.8
aiohttp==2.1.0
requests==2.17.0
hug==2.3.0
